Moved widget code segments to make overall code readable; Changed branch name
