+++ lib/scenes/details_page.dart
@@ -1,6 +1,8 @@
 import 'dart:io';
 
 import 'package:cannoli_app/comparison_graph.dart';
+import 'package:cannoli_app/widgets/details_graph_view.dart';
+import 'package:cannoli_app/widgets/emission_chart.dart';
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_staggered_grid_view/flutter_staggered_grid_view.dart';
@@ -172,13 +174,13 @@ class _DetailsPageState extends State<DetailsPage> {
   Widget getWidgetGroup() {
     switch (widgetSelection) {
       case WidgetList.generalGraph:
-        return getGeneralWidgetGroup();
+        return getAllGraphWidgets();
       case WidgetList.transportGraph:
-        return getTransportWidgetGroup();
+        return getTransportGraphWidgets();
       case WidgetList.homeGraph:
-        return getHomeWidgetGroup();
+        return getHomeGraphWidgets();
     }
-    return getGeneralWidgetGroup();
+    return getAllGraphWidgets();
   }
 
   @override
@@ -200,244 +202,8 @@ class _DetailsPageState extends State<DetailsPage> {
   }
 }
 
-Widget mychart1Items(String title, String priceVal, String subtitle) {
-  return Material(
-    color: Colors.white,
-    elevation: 1.0,
-    borderRadius: BorderRadius.circular(24.0),
-    child: Center(
-      child: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    title,
-                    style: TextStyle(
-                      fontSize: 20.0,
-                      color: Colors.blueAccent,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    priceVal,
-                    style: TextStyle(
-                      fontSize: 30.0,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    subtitle,
-                    style: TextStyle(
-                      fontSize: 20.0,
-                      color: Colors.blueGrey,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: new Sparkline(
-                    data: [
-                      0.0,
-                      1.0,
-                      1.5,
-                      2.0,
-                      0.0,
-                      0.0,
-                      -0.5,
-                      -1.0,
-                      -0.5,
-                      0.0,
-                      0.0
-                    ],
-                    lineColor: Color(0xffff6101),
-                    pointsMode: PointsMode.all,
-                    pointSize: 8.0,
-                  ),
-                ),
-              ],
-            )
-          ],
-        ),
-      ),
-    ),
-  );
-}
-
-Widget comparisonCard(String title, String subTitle) {
-  return Material(
-    color: Colors.white,
-    elevation: 1.0,
-    borderRadius: BorderRadius.circular(24.0),
-    child: Center(
-      child: Padding(
-        padding: EdgeInsets.all(8.0),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    title,
-                    style: TextStyle(
-                      fontSize: 20.0,
-                      color: Colors.blueAccent,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    subTitle,
-                    style: TextStyle(
-                      fontSize: 20.0,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(6),
-                ),
-                Padding(padding: EdgeInsets.all(1.0), child: ComparisonGraph()),
-              ],
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
-}
-
-List<CircularStackEntry> circularData = <CircularStackEntry>[
-  new CircularStackEntry(
-    <CircularSegmentEntry>[
-      new CircularSegmentEntry(700.0, Color(0xff4285F4), rankKey: 'Q1'),
-      new CircularSegmentEntry(1000.0, Color(0xfff3af00), rankKey: 'Q2'),
-      new CircularSegmentEntry(1800.0, Color(0xffec3337), rankKey: 'Q3'),
-      new CircularSegmentEntry(1000.0, Color(0xff40b24b), rankKey: 'Q4'),
-    ],
-    rankKey: 'Quarterly Profits',
-  ),
-];
-
-Material myCircularItems(String title, String subtitle) {
-  return Material(
-    color: Colors.white,
-    elevation: 1.0,
-    borderRadius: BorderRadius.circular(24.0),
-    child: Center(
-      child: Padding(
-        padding: EdgeInsets.all(8.0),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: <Widget>[
-                Padding(
-                  padding: EdgeInsets.all(3.0),
-                  child: Text(
-                    title,
-                    style: TextStyle(
-                      fontSize: 20.0,
-                      color: Colors.blueAccent,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(1.0),
-                  child: Text(
-                    subtitle,
-                    style: TextStyle(
-                      fontSize: 10.0,
-                    ),
-                  ),
-                ),
-                Padding(
-                  padding: EdgeInsets.all(3.0),
-                  child: AnimatedCircularChart(
-                    size: const Size(150.0, 150.0),
-                    initialChartData: circularData,
-                    chartType: CircularChartType.Pie,
-                  ),
-                ),
-              ],
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
-}
-
-Widget getTransportWidgetGroup() {
-  return Container(
-    padding: EdgeInsets.only(top: 8.0),
-    child: Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        // Page Layout Title Text Box
-        pageLayoutTextBoxWidget('Transport: Overview', 18.0),
-      ],
-    ),
-  );
-}
-
-Widget getHomeWidgetGroup() {
-  return Container(
-    padding: EdgeInsets.only(top: 8.0),
-    child: Row(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: <Widget>[
-        // Page Layout Title Text Box
-        pageLayoutTextBoxWidget('Home: Overview', 18.0),
-      ],
-    ),
-  );
-}
-
-Widget getGeneralWidgetGroup() {
-  return Container(
-      padding: EdgeInsets.only(top: 8.0),
-      child: Column(
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              // Page Layout Title Text Box
-              pageLayoutTextBoxWidget('All Emissions: Overview', 18.0),
-            ],
-          ),
-          Container(
-            padding: EdgeInsets.only(top: 8.0),
-            child: myCircularItems("Today's Total Consumption", "250kg CO2"),
-          ),
-          Container(
-            padding: EdgeInsets.only(top: 8.0),
-            child:
-                mychart1Items("Home energy Consumption", "Weekly", "300kg CO2"),
-          ),
-          Container(
-            padding: EdgeInsets.only(top: 8.0),
-            child: comparisonCard("Comparison with ideal emission", "Weekly"),
-          ),
-        ],
-      ));
-}
-
 /*
-Widget getGeneralWidgetGroup() {
+Widget getAllGraphWidgets() {
   return ListView(
     children: <Widget>[
       Container(
